Bump package installation timeout for "docker git" due to slow install
As seen recently in
https://openqa.opensuse.org/tests/1778520#step/openqa_webui/6
package downloads and installations can timeout, likely due to
performance degradations in the infrastructure. Also, the list of
package dependencies that are installed along with docker and git can be
heavy including regeneration of initrd images hence it makes sense to
bump the timeout for this specific installation call.

Related progress issue: https://progress.opensuse.org/issues/93826