Prevent packagekit interfering with later installations
We already disabled the packagekit service during installation of the
openQA webUI but we enabled it again at the end of
tests/install/openqa_webui which means it could interfer with
installations in a later module. As we do not need or actually do not
want any updates interfering within one test run we should just keep
packagekit disabled during the runtime of the test. Hence this commit
removes the re-enablement of the packagekit service.

Additionally I changed the call to the simpler, consistent "mask --now"
call as this is supported in all current openSUSE versions. Also
changing to use 'assert_script_run' to follow the best practice of using
"assert_script_run" over "script_run".