Load python test modules by default with new variable
Instead of needing to explicitly specify python in the job templates we
should load python tests by default with a more explanatory variable
name while still offering to skip on problems.

This is inspired by
https://github.com/os-autoinst/os-autoinst-distri-openQA/pull/114/files#diff-093d7aec968469f705e116051d4161bd82c22f4add7145791e04e46101881fecR44
showing me that job templates need the explicit variable when it
shouldn't be necessary to reiterate that.