Test setup-multi-machine by calling it as part of worker setup
This just adds a call to "setup-multi-machine" as part of the worker
setup. With that we ensure that the command be successfully executed and
that it does not have a detrimental impact on the executed test steps
following. However it does not ensure that multi-machine tests would
actually work based on that.

Related progress issue: https://progress.opensuse.org/issues/133025