Allow the post_fail_hook ip command to fail
It may fail in case this IP is not actually set on the bridge.
Also the ctrl-c needs to happen before, because if the

    grep ... tail -f -n0 nohup.out

fails, it's not killed. Since ctrl-c shouldn't do any harm, I guess we can
do it regardless of which test we are running.

This ensures that the post_fail_hook can be completed.