Fix display of logfile 'cpuinfo'
A browser on the openQA web UI views can directly display files with
extension .txt so we should upload cpuinfo output as such. This fixes
the problem that the cpuinfo file as uploaded by the post_fail_hook
needs to be downloaded and explicitly opened as text file whereas it
could be viewed directly in the browser.