Record results of failing Minion jobs
This might help to debug https://progress.opensuse.org/issues/163784 where
the Git output is omitted by the journal (as it contains non-printable
characters). This way we'll also always see the concrete error message why
a Minion job eventually failed (and not just errors logged during the
execution of the job).