Also upload the whole journal to find other errors
Recently we observed a problem with the os-autoinst-openvswitch service.
While the relevant system journal content was shown in screenshots it
was not uploaded in an easy to process log file so I am adding another
call to also upload the complete system journal which should not be that
big on those clean test OS installations.