Implement module to validate Autoyast template profiles
With the use of the test variable AUTOYAST we can pass a erb profile tfor
validation. The module exercises the happy path. Meaning that it checks only
one screen which it comes at the end of the validation of `yast2 autoyast
check-profile`. For any warning or error, yast will throw a popup and the test
should fail.

https://progress.opensuse.org/issues/156169

Signed-off-by: ybonatakis <ybonatakis@suse.com>