Ensure test_running-testresults.tar.gz contains image files
files in testresults.tar.gz are not opening, likely because are symbolic
links. Ensure that the tar contains the actual files and can be accessed.
For that tar takes `-h` option which comes with some caveats. As such the
assert_script_run is replaced by script_run. The reason is that it can fail
due to broken links or circular references.

https://progress.opensuse.org/issues/169249

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>