Fix adjusting firefox config
"prepare_firefox_autoconfig" needs a text terminal to type into. It
seems at some time in the past a regression was introduced that the
firefox config is not typed into a text terminal anymore after any
previous test module did not end up on the text terminal anymore.

This fixes the entering of configuration by explicitly switching to the
root console. This should fix the sporadic issue of firefox asking in a
first-time wizard of confirmation instead of showing the selected page
that the openQA test expects.

Related progress issue: https://progress.opensuse.org/issues/181187