Verify that worker register to the openqa_webui container
Even with a proper client.conf the worker was unable to register with
the webui. The redirection to the localhost is bypassed when the
worker provides WORKER_HOSTNAME, so it was not the main problem. The
webui container in the other side runs reserve proxy, however is the
template. Overall the solution was a sync of the configuration files
with the expected openqa host. To do that, specific files in the host
are overriden instead of manipulated or created new ones. That was to
keep the original files in the repo intact as it is likely already
used as is in other cases which it is not known at that point.
Main changes is to use Fake for the authentication and then override
the ENTRYPOINT of the worker to request a /login before run the entry
command.

Other possible solution might be to edit the Dockerfile or maybe do a
edit the db.

Note that the containers now use the `--hostname` which it doesn't
make any difference but it is nice to have the hostname set in the
containers.

poo: https://progress.opensuse.org/issues/186651
Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>