Update Docker build arguments and directories
The templates are almost entirely modified and in order to provide a clean
configuration for the test case, it is better to not use them at all. This,
hopefully descreases the complexity and provides a clean overview of the
minimal configuration that the container is needed.
Furthermore, this commit is depended on the webui Dockerfile, which should be
provide the proxy configuration on the webui container for the build.

Depends-on: https://github.com/os-autoinst/openQA/pull/6751
Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>