Make directories available to container under SELinux
The "z" flag marks this file as being used in multiple containers.
https://docs.docker.com/engine/storage/bind-mounts/#configure-the-selinux-label

`chcon -t container_file_t` will mark a file to be allowed by a container.

I used the z flag for single files and chcon for the directories that are used
in multiple tests/containers.

Issue: https://progress.opensuse.org/issues/191809